requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-filter==25.2
django-cors-headers==4.9.0

gunicorn==25.1.0
whitenoise==6.12.0

pytest==9.0.2
pytest-django==4.12.0
pytest-cov==7.0.0